chore(doc-engine): introduce top-level doc-engine package with shims
