+++ calculogic-doc-engine/README.md
@@ -1,0 +1,12 @@
+# @calculogic/doc-engine
+
+Doc-engine core package extracted from app; app currently consumes via shims in `src/doc-engine/*`.
+
+## Public exports
+
+- `ContentProviderRegistry`, `parseContentRef`, `splitNamespace`
+- Doc-engine resolution and provider types
+
+## Usage note
+
+Do not deep-import internal files; use the package root export.
