+++ calculogic-doc-engine/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@calculogic/doc-engine",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  }
+}
