+++ doc/ConventionRoutines/CCPP.md
@@ -1,0 +1,11 @@
+# CCPP (Concern/Code Provenance Policy)
+
+## Purpose
+Define lightweight provenance and comment rules for Calculogic repositories.
+
+## Rules
+1. Keep comments focused on intent, constraints, and invariants.
+2. Avoid stale implementation comments; prefer self-descriptive names and tests.
+3. When behavior is governed by architecture/docs, reference the source doc in nearby comments or commit/PR notes.
+4. Preserve public behavior unless a task explicitly changes it.
+5. Keep changes minimal and scoped to the concern being modified.
