+++ package.json
@@ -8,7 +8,7 @@
     "build": "tsc -b && vite build",
     "lint": "eslint .",
     "preview": "vite preview",
-    "test": "node --test --experimental-strip-types test/**/*.test.mjs calculogic-validator/test/**/*.test.mjs",
+    "test": "node --test --experimental-strip-types test/**/*.test.mjs calculogic-doc-engine/test/**/*.test.mjs calculogic-validator/test/**/*.test.mjs",
     "validate:naming": "node --experimental-strip-types calculogic-validator/scripts/validate-naming.mjs",
     "validate:all": "node --experimental-strip-types calculogic-validator/scripts/validate-all.mjs",
     "health:validator": "node --experimental-strip-types calculogic-validator/scripts/validator-health-check.host.mjs",
@@ -48,6 +48,7 @@
     "typescript": "~5.7.2",
     "typescript-eslint": "^8.26.1",
     "vite": "^6.3.1",
+    "@calculogic/doc-engine": "file:calculogic-doc-engine",
     "@calculogic/report-capture": "file:calculogic-validator/tools/report-capture",
     "@calculogic/validator": "file:calculogic-validator"
   }
