+++ src/doc-engine/index.ts
@@ -1,14 +1,2 @@
-import { ContentProviderRegistry, parseContentRef, splitNamespace } from './registry.ts';
-
-export { ContentProviderRegistry, parseContentRef, splitNamespace };
-export type {
-  ContentProvider,
-  ContentResolutionRequest,
-  ContentResolutionResult,
-  FoundContent,
-  InvalidRef,
-  MissingContent,
-  NoProvider,
-  NotFound,
-  ParsedContentRef,
-} from './types.ts';
+export { ContentProviderRegistry, parseContentRef, splitNamespace } from '@calculogic/doc-engine';
+export type * from '@calculogic/doc-engine';
